requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pathlib==1.0.1
psycopg2==2.9.3
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
gdal==3.3.3
numpy==1.22.4
pyjq==2.5.2